requirements.txt
# Core LangChain and LangGraph
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0

# Azure OpenAI Support
openai>=1.0.0

# Data validation and settings
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Environment management
python-dotenv>=1.0.0

# Utilities
typing-extensions>=4.0.0
colorama>=0.4.6

# Optional: Web Interface (if needed later)
fastapi>=0.100.0
uvicorn>=0.23.0

# Testing
pytest>=7.0.0
